feat(ui): overflow: scroll -> auto
I don't have a strong opinion on overflow: scroll vs auto, but the vast
majority of our components use auto, and I think it's important for it to be
consistent across the product unless there's a particularly good reason
otherwise.

Note that two instances of overflow: scroll are not changed since they are
being used to measure scrollbar size, which is arguably dubious but will
obviously break if directly removed.