feat(perf): Initial pass of generic performance widget
This adds a generic performance widget which takes in a chart and a query and displays something akin to our existing landing panels. This supports just basic histograms at the moment, and the ability to switch between them, with your preference being stored in local storage.

Also includes a small change that was dropped to actually start using the ops breakdown filter context, as well as adding a page-wide error context so setError does not need to be drilled.