feat(py): Support frontend manifest lookups from `sentry.options`
Previously, the webpack manifest was expected to be on the filesystem. This PR allows the usage of a manifest that is saved to `sentry.options`. This allows us to decouple the loading of frontend assets from the backend.

This falls back to the manifest located on the filesystem if nothing was found in options.